requirements.txt
# --- Transformers & LoRA ---
transformers>=4.45.0
accelerate>=0.34.0
huggingface-hub
peft>=0.12.0
bitsandbytes>=0.43.3
qwen-vl-utils>=0.0.8

# --- Utilities ---
numpy<2.0.0
pandas
scikit-learn
pillow
tqdm
tensorboard
opencv-python

# --- Low-level Dependencies ---
pytz
python-dateutil
six
typing_extensions
requests
packaging
pyyaml